feat: alterações para exibir produtos sem opção de favoritos para usuarios deslogados

+++ src/hooks/useRemoveFavoriteMutate.tsx
@@ -1,25 +1,28 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import axios, { AxiosPromise } from "axios";
+import { LoginResponse } from "../interfaces/login-data.interface";
 
 const { VITE_API_URL } = import.meta.env;
 
-const removeFav = async (id: string, favId: string): AxiosPromise<void> => {
-  const token = localStorage.getItem('token')?.replace(/^"(.*)"$/, '$1');
+const removeFav = async (id: string, favId: string, token: LoginResponse | null): Promise<AxiosPromise<void> | null> => {
+  if (token) {
+    const response = await axios.delete<void>(`${VITE_API_URL}/favorite/${id}/${favId}`, {
+      headers: {
+        Authorization: `Bearer ${token}`
+      }
+    });
+  
+    return response;
+  }
 
-  const response = await axios.delete<void>(`${VITE_API_URL}/favorite/${id}/${favId}`, {
-    headers: {
-      Authorization: `Bearer ${token}`
-    }
-  });
-
-  return response;
+  return null;
 };
 
-const useRemoveFavoriteMutate = (id: string, favId: string) => {
+const useRemoveFavoriteMutate = (id: string, favId: string, token: LoginResponse | null) => {
   const queryClient = useQueryClient();
 
   const mutate = useMutation({
-    mutationFn: () => removeFav(id, favId),
+    mutationFn: () => removeFav(id, favId, token),
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ['favorite-data'] });
     }
